rename constants file to index
